#hotfix corrected the data type when assigning the image size, deleted unused exceptions

+++ src/main/java/ru/skypro/homework/service/impl/ImageServiceImpl.java
@@ -51,7 +51,7 @@ public class ImageServiceImpl implements ImageService {
 
         try {
             image.setPath(path)
-                    .setSize(file.getSize())
+                    .setSize((int) file.getSize())
                     .setMediaType(file.getContentType())
                     .setData(file.getBytes());
         } catch (IOException e) {
